+++ web/src/main/java/com/isa/domain/api/EventApi.java
@@ -25,6 +25,11 @@ public class EventApi {
     private WeblinkExternal weblinkExternal;
 
 
+    @JsonProperty("schedule")
+    @JsonIgnore
+    private String schedule;
+
+
     @JsonProperty("attachments")
     @JsonIgnoreProperties(ignoreUnknown = true)
     private List<AttachmentExternal> attachments;
