+++ web/src/main/java/com/isa/servlet/ListOfEvents.java
@@ -2,7 +2,7 @@ package com.isa.servlet;
 
 import com.isa.config.TemplateProvider;
 import com.isa.domain.dto.EventDto;
-import com.isa.service.EventService;
+import com.isa.service.service.domain.EventService;
 import freemarker.template.Template;
 import freemarker.template.TemplateException;
 import org.slf4j.Logger;
